Have to force bbox in the yt bounded box frontend to be float32 at the moment. Otherwise roundoff in filter_bbox doesn't mask out all the particles depending on the particular precision, and a runtime error in the yt octtree contstruction is thrown.  Also try to match rmin/rmax logic with midx code.
--HG--
branch : yt-3.0

+++ yt/frontends/sdf/data_structures.py
@@ -74,7 +74,7 @@ class SDFDataset(Dataset):
         self.n_ref = n_ref
         self.over_refine_factor = over_refine_factor
         if bounding_box is not None:
-            bbox = np.array(bounding_box, dtype="float64")
+            bbox = np.array(bounding_box, dtype="float32")
             if bbox.shape == (2, 3):
                 bbox = bbox.transpose()
             self.domain_left_edge = bbox[:,0]
